feat(News): initial typing
