+++ src/components/view/news/news.ts
@@ -1,0 +1,42 @@
+import { ArticlesType } from 'components/controller/types';
+import './news.css';
+
+class News {
+    draw(data: ArticlesType['articles']) {
+        const news = data.length >= 10 ? data.filter((_item, idx) => idx < 10) : data;
+
+        const fragment = document.createDocumentFragment();
+        const newsItemTemp = document.getElementById('newsItemTemp') as HTMLTemplateElement;
+
+        news.forEach((item, idx) => {
+            const newsClone = newsItemTemp.content.cloneNode(true) as HTMLElement;
+
+            if (idx % 2) {
+                (newsClone.querySelector('.news__item') as HTMLElement).classList.add('alt');
+            }
+
+            (newsClone.querySelector('.news__meta-photo') as HTMLElement).style.backgroundImage = `url(${
+                item.urlToImage || 'img/news_placeholder.jpg'
+            })`;
+            (newsClone.querySelector('.news__meta-author') as HTMLElement).textContent =
+                item.author || item.source.name;
+            (newsClone.querySelector('.news__meta-date') as HTMLElement).textContent = item.publishedAt
+                .slice(0, 10)
+                .split('-')
+                .reverse()
+                .join('-');
+
+            (newsClone.querySelector('.news__description-title') as HTMLElement).textContent = item.title;
+            (newsClone.querySelector('.news__description-source') as HTMLElement).textContent = item.source.name;
+            (newsClone.querySelector('.news__description-content') as HTMLElement).textContent = item.description;
+            (newsClone.querySelector('.news__read-more a') as HTMLElement).setAttribute('href', item.url);
+
+            fragment.append(newsClone);
+        });
+
+        document.getElementsByClassName('news')[0].innerHTML = '';
+        document.getElementsByClassName('news')[0].appendChild(fragment);
+    }
+}
+
+export default News;
